Split config.c into a separate object file
